Resolve review comments about backup compontent refactoring
