+++ app/src/main/java/com/stevesoltys/backup/transport/component/provider/ContentProviderBackupConstants.java
@@ -3,7 +3,14 @@ package com.stevesoltys.backup.transport.component.provider;
 /**
  * @author Steve Soltys
  */
-class ContentProviderBackupConstants {
+public interface ContentProviderBackupConstants {
 
-    static final String SALT_FILE_PATH = "salt";
+    String SALT_FILE_PATH = "salt";
+
+    String DEFAULT_FULL_BACKUP_DIRECTORY = "full/";
+
+    String DEFAULT_INCREMENTAL_BACKUP_DIRECTORY = "incr/";
+
+    long DEFAULT_BACKUP_QUOTA = Long.MAX_VALUE;
+
 }
